+++ lib/lotus/web/cell_formatter.ex
@@ -1,6 +1,8 @@
 defmodule Lotus.Web.CellFormatter do
   @moduledoc "Formats arbitrary DB values for safe display in result tables."
 
+  # def format(v), do: Lotus.Value.to_display_string(v)
+
   def format(nil), do: ""
 
   # Dates & times