+++ lib/lotus/web/components/icons.ex
@@ -161,6 +161,29 @@ defmodule Lotus.Web.Components.Icons do
     """
   end
 
+  attr(:rest, :global)
+
+  def layers(assigns) do
+    ~H"""
+    <.svg_outline {@rest}>
+      <path d="M12.83 2.18a2 2 0 0 0-1.66 0L2.6 6.08a1 1 0 0 0 0 1.83l8.58 3.91a2 2 0 0 0 1.66 0l8.58-3.9a1 1 0 0 0 0-1.83z"/>
+      <path d="M2 12a1 1 0 0 0 .58.91l8.6 3.91a2 2 0 0 0 1.65 0l8.58-3.9A1 1 0 0 0 22 12"/>
+      <path d="M2 17a1 1 0 0 0 .58.91l8.6 3.91a2 2 0 0 0 1.65 0l8.58-3.9A1 1 0 0 0 22 17"/>
+    </.svg_outline>
+    """
+  end
+
+  attr(:rest, :global)
+
+  def search(assigns) do
+    ~H"""
+    <.svg_outline {@rest}>
+      <path d="m21 21-4.34-4.34"/>
+      <circle cx="11" cy="11" r="8"/>
+    </.svg_outline>
+    """
+  end
+
   # Column type icons for schema explorer
   attr(:rest, :global)